A bit of refactor to utils
